requirements.txt
customtkinter
playwright
python-vlc
requests
Pillow
